Final Push
Attempted to implement the usage of the Authenticator API to authenticate users however could not find a way to do so. I have left the file 'authenticator' as proof that an attempy was made but had to revert the code made in other files to ensure functionality. Added the 'Dockerfile' to share my code on the Docker Hub.